pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dmas"
version = "0.2.0"
description = "Designing Multi-Agent Systems — book companion code"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0",
    "requests>=2.31",
]

[project.optional-dependencies]
autogen = ["autogen-agentchat>=0.7", "autogen-ext[openai]>=0.7"]
agent-framework = ["agent-framework>=1.0.0rc1"]
langgraph = ["langgraph>=0.2", "langchain-openai>=0.3"]
all = ["dmas[autogen]", "dmas[agent-framework]", "dmas[langgraph]"]

[project.scripts]
dmas-ch1 = "dmas.ch1.main:cli"
dmas-ch4 = "dmas.ch4.main:cli"

[tool.setuptools.packages.find]
where = ["src"]
